Remove unnecessary type cast
git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14770 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/gui/editors/segment/TrackButtons.cpp
@@ -928,7 +928,7 @@ TrackButtons::selectInstrument(Track *track, Instrument *instrument)
         // instrument, so they're no longer ready (making them
         // ready is done just-in-time elsewhere), nor is thru
         // channel ready.
-        if (((int)segment->getTrack()) == trackId)
+        if (segment->getTrack() == trackId)
             sequenceManager->segmentInstrumentChanged(segment);
 
     }
